requirements.txt
django
djangorestframework
markdown
django-filter
rest_framework_simplejwt
drf-yasg
